Fixed the case where no testUser.json file exists previously

+++ test/lib/generateTestUser.test.js
@@ -17,7 +17,14 @@ describe('generateTestUser()', function () {
 
     beforeEach(function (done) {
 
-        fs.unlink(expectedFilePath, done);
+        fs.exists(expectedFilePath, function (exists) {
+
+            if (!exists) { return done(); }
+
+            fs.unlink(expectedFilePath, done);
+
+        });
+
 
     });
